started changing stuff
