+++ client/navbar/navbar.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const NavBar: React.FC = () => {
   return (
-    <nav>
+    <nav className="navbar">
       &nbsp;
     </nav>
   )
